create timer and fixed time
