+++ package.json
@@ -49,6 +49,7 @@
     "swiper": "^7.4.1",
     "vue": "^2.6.14",
     "vue-loader": "^15.9.8",
+    "vue-moment": "^4.1.0",
     "vue-template-compiler": "^2.6.14"
   }
 }
