+++ src/index.html
@@ -4,9 +4,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://fonts.googleapis.com/css2?family=PT+Sans&display=swap" rel="stylesheet">
+    
     <!-- SEO tags -->
     <title>Главная страница &bull; Webpack Starter</title>
-    <script src="https://code.iconify.design/2/2.1.0/iconify.min.js"></script>
     <meta name="description" content="">
     <meta property="og:title" content="">
     <meta property="og:image" content="">
@@ -23,6 +23,9 @@
     <meta name="msapplication-TileColor" content="#000000">
     <meta name="msapplication-config" content="/img/favicons/browserconfig.xml">
     <meta name="theme-color" content="#000000">
+
+    <!-- JS libs -->
+    <script src="https://code.iconify.design/2/2.1.0/iconify.min.js"></script>
 
     <!-- Main styles -->
     <link rel="stylesheet" href="/css/app.css?v=1.0.0">
