+++ src/js/components/MatchesFilter.vue
@@ -39,12 +39,5 @@
         this.$emit('input', this.usersInput);
       },
     },
-    mounted(){},
-    methods: {
-
-    }
   }
-  
-
-
 </script>
