+++ src/styles/app.scss
@@ -10,3 +10,4 @@
 @import "components/search";
 @import "components/table";
 @import "components/versus";
+@import "components/keyframe";
