+++ src/styles/components/_loading.scss
@@ -1,9 +1,10 @@
 .loading{
+  margin-top: 59px;
   width: 100%;
   text-align: center;
 
   &__text{
-    font-size: 24px;
+    font-size: rem(24px);
     font-weight: bold;
     margin-top: 14.5px;
     margin-bottom: 20px;      //to avoid text cutting
@@ -11,15 +12,12 @@
 
   .iconify{
     color: $primary;
-    font-size: 63px;
+    font-size: rem(64px);
     animation: spin 2s linear 0s infinite;
   }
   
 }
 
-@keyframes spin {
-  from {transform:rotate(0deg);}
-  to {transform:rotate(360deg);}
-}
+
