+++ src/styles/components/_table.scss
@@ -1,4 +1,4 @@
-.table{
+.match{
   min-width: 740px;
   position: relative;
   height: 80px;
@@ -6,13 +6,11 @@
   background-color: $bg-white;
   display: flex;
   color: black;
-  font-size: 14px;
+  font-size: rem(14px);
 
   &__time, &__translation{
     color: gray;
-    margin: auto;
-    margin-left: 15px;
-    margin-right: 0px;
+    margin: auto 0 auto 15px;
   }
 
   &__translation{
